Support nonstandard attributes
Assigning a property to an element directly sets the
property as an attribute if it is a standard attribute
like "draggable". However, this is not the case for custom
attributes like "data-test".

By uniformly using element.setAttribute, the custom attributes
get inserted as DOM attributes as well.

Similarly, we can setAttribute("class") when the JSX key is
"className", to avoid the (el as any) type assertion.