Add special handling for SVG elements
Svg elements need to be created with document.createElementNS.
Creating svg elements in the default HTML namespace results in them e.g.
not having widths and heights in the DOM.

Also defaulting to setting all svg props with setAttribute.
For example, directly assigning to the "cx" prop of a circle
element throws because it only has a getter.